move preload into proper directory
