+++ .eslintrc.json
@@ -7,6 +7,7 @@
   "ignorePatterns": [
     "forge.config.ts",
     "vite.*.config.*",
+    "out/", // build output
     "stories/" // exclude stories from linting for now
   ],
   "extends": [
